frontend: move component/fiat to routes/settings
